Extract "from" tests into separate test class
